+++ KFParticle/KFParticlePVReconstructor.h
@@ -41,13 +41,13 @@ class KFParticlePVReconstructor{
   int NPrimaryVertices() const { return fPrimVertices.size(); }
   KFParticle &GetPrimVertex(int iPV=0)   { return fPrimVertices[iPV]; };
   KFVertex   &GetPrimKFVertex(int iPV=0)   { return fPrimVertices[iPV]; };
-  vector<short int>& GetPVTrackIndexArray(int iPV=0) { return fClusters[iPV].fTracks; }
+  vector<int>& GetPVTrackIndexArray(int iPV=0) { return fClusters[iPV].fTracks; }
   KFParticle &GetParticle(int i){ assert( i < fNParticles );          return fParticles[i];    };
   
   void SetBeamLine(KFParticle& p) { fBeamLine = p; fIsBeamLine = 1; }
   bool IsBeamLine() const { return fIsBeamLine; }
   
-  void AddPV(const KFVertex &pv, const vector<short int> &tracks);
+  void AddPV(const KFVertex &pv, const vector<int> &tracks);
   void AddPV(const KFVertex &pv);
   void CleanPV() { fClusters.clear(); fPrimVertices.clear(); }
 
@@ -67,7 +67,7 @@ class KFParticlePVReconstructor{
   
   struct KFParticleCluster {
     KFParticleCluster():fTracks(0) {};
-    vector<short int> fTracks;
+    vector<int> fTracks;
     float fP[3];
     float fC[6];
   };
